refactor: 패키지명 변경 및 util 클래스 생성자 제거
